move the item-data interface to a specific model file
