+++ src/app/items/containers/item-children-edit-form/item-children-edit-form.component.ts
@@ -1,5 +1,5 @@
 import { Component, Input, OnDestroy, OnInit, ViewChild } from '@angular/core';
-import { ItemData } from '../../services/item-datasource.service';
+import { ItemData } from '../../models/item-data';
 import {
   ChildDataWithId,
   hasId,
