+++ src/app/items/containers/item-edit-wrapper/item-edit-wrapper.component.ts
@@ -1,5 +1,5 @@
 import { Component, Input, OnChanges, OnDestroy, OnInit } from '@angular/core';
-import { ItemData } from '../../services/item-datasource.service';
+import { ItemData } from '../../models/item-data';
 import {
   AbstractControl,
   UntypedFormBuilder,
