+++ src/app/items/containers/item-header/item-header.component.ts
@@ -1,5 +1,5 @@
 import { Component, Input } from '@angular/core';
-import { ItemData } from '../../services/item-datasource.service';
+import { ItemData } from '../../models/item-data';
 import { NgIf } from '@angular/common';
 
 @Component({
