+++ src/app/items/containers/item-permissions/item-permissions.component.ts
@@ -1,5 +1,5 @@
 import { Component, EventEmitter, Input, OnChanges, Output } from '@angular/core';
-import { ItemData } from '../../services/item-datasource.service';
+import { ItemData } from '../../models/item-data';
 import {
   ProgressSelectValue,
   ProgressSelectComponent,
