+++ src/app/items/containers/item-unlock-access/item-unlock-access.component.ts
@@ -1,5 +1,5 @@
 import { Component, ElementRef, Input, OnChanges, OnDestroy, QueryList, ViewChild, ViewChildren } from '@angular/core';
-import { ItemData } from '../../services/item-datasource.service';
+import { ItemData } from '../../models/item-data';
 import { OverlayPanel, OverlayPanelModule } from 'primeng/overlaypanel';
 import { BehaviorSubject, debounceTime, merge, ReplaySubject, Subject, switchMap } from 'rxjs';
 import { distinctUntilChanged, filter, map, share, shareReplay } from 'rxjs/operators';
