+++ src/app/items/containers/sub-skills/sub-skills.component.ts
@@ -4,7 +4,7 @@ import { distinctUntilChanged, map, switchMap } from 'rxjs/operators';
 import { bestAttemptFromResults } from 'src/app/items/models/attempts';
 import { isASkill } from 'src/app/items/models/item-type';
 import { GetItemChildrenService } from '../../../data-access/get-item-children.service';
-import { ItemData } from '../../services/item-datasource.service';
+import { ItemData } from '../../models/item-data';
 import { mapToFetchState } from 'src/app/utils/operators/state';
 import { canCurrentUserViewContent } from 'src/app/items/models/item-view-permission';
 import { ItemChildrenListComponent } from '../item-children-list/item-children-list.component';
